Added new routines to storage.f90:
- storage_copy_explicit
- storage_copy_explicit2D

In many situations it is useful to copy only parts of 
an array to another array. It is even possible that both
arrays have different length provided the subarray to be
copied from the source array fits into the destination array.

In addition to the standard storage_copy routine, the new 
routines provide the input parameters
istart_source, istart_dest and ilength which denote the offset of
the source and destination arrays and the length of the subarray
to be copied.

Example

integer a(10),b(40)

To copy a(3:13)=B(25:35) use
storage_copy(h_b,h_a,25,3,10)



git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@226 4b12f95d-9313-0410-afc9-bd26427041f1