1. Fixed an ugly bug imported from Feast: "SP was not SP but DP". The kind definitions in fsystem.f90 wre wrong such that SP=8 and DP=8. As a consequence, some Blas routines for single arrays produced wrong results.
One of the copyVector routines had to be modified.

!!! ATTENTION !!!
It may be possible, that some of the applications do not compile
because the compiler complains about type mismatch ;-)


2. Added a new routine to storage.f90

- storage_realloc

An existing array can be reallocated at a different size. All quantities
like type, dimension, name will be the same. In addition, the handle
does not change and can be used without problem. However, arrays which
are associated to the handle by means of the storage_getbase_XXX routines
will no longer be valid.



git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@229 4b12f95d-9313-0410-afc9-bd26427041f1