Fixed some bugs in subroutine lsyssc_scalarMatVec if the matrix is stored in interleave format. Now, matrix multiplication does work.
Fixed some small bug in module linearsystemblock. If a scalar
matrix/vector that is stored in interleave format is converted to its
block counterpart then NEQ and NCOLS are set to NEQ*NVAR and
NCOLS*NVAR respectively. As a consequence, the block objects are still
compatible to each other. Moreover, if one of these objects is used to
create a new on, e.g., a 1-block vector is used as template to create
another block vector, then the correct size of NEQ*NVAR vector entires
is used.



git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@613 4b12f95d-9313-0410-afc9-bd26427041f1