Added HDF5 support to the Featflow kernel. If applications are compiled as before, HDF5 support is not enabled. In order to enable HDF5 support you must add HDF5=YES to your make command. Then the required hdf libraries (external source) are generated and some additional module files are included in the build process.
HDF5 can be used to store and retrieve large data to and from the filesystem. The h5lite module is designed to simplify the use of HDF5. To begin with, you create a h5file object that connects the physical file to some derived data type. You can set/get attributes and data sets to and from the file. At the moment, the routine provide the basic operations. I future, I will include some more checks, whether the file is open, etc. Their interfaces will not change (hopyfully), so you may start to use them.

The linearsystemh5io module provides all routine that are required to store scalar/block vectors and matrices in HDF5. As before, some additional checks may be included but the routine should be already working.



git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@1410 4b12f95d-9313-0410-afc9-bd26427041f1