Created new directory for some standard and nonstandard data structures. To be more precise, this directory contains modules for lists, trees, graphs, arraylists and quadtrees. (@Michael. I removed the quadtree from directory System and placed it in here. Can you please update your Windows build environment. Perhaps we should also move the stack from the System directory into this directory.)
Most of the data structures are self-explaining. An arraylist is an array of lists arrays (as the name suggests). The idea is to have immediate access to the array position and then have an ordered/unordered list of data associated to the array. Think of Kld and Kcol or of hash data structures.

A final word about the implementation: All modules are internally realized as 1D and 2D arrays (of cause making use of the storage module) in order to provide best performance. Some year ago I implemented the tree module (which indeed is an AVL tree) making extensive use of pointers and derived type for each node. This is, what you learn in any lecture about modern, object-oriented programming. It works but is not competitive at all ;-) At least not in Fortran! The modules may look ugly, but they are extremely fast and, moreover, most of the code already exists some time and is tested quite well (hopefully). Matthias

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@1620 4b12f95d-9313-0410-afc9-bd26427041f1