Tried to illustrate refinement/conversion routines by improving the ASCII-arts.
Added elements-meeting-at-vertex data structure to the adaptivity structure. This auxiliary structure is used in the recoarsening routine in order to update all neighbouring elements of a deleted/renumbered vertex. The elements-at-vertex structure is generate by the hadapt_genElementsAtVertex subroutine. During refinement/conversion it is updated continuously so as to reflect the current grid. The recoarsening algorithm is still to be implemented (I will do, soon!).

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@1745 4b12f95d-9313-0410-afc9-bd26427041f1