New features        : - Added implementation of subdomains. Problems solved     : Bugzilla-Bug-ID     : Additional comments :
A subdomain can be generated by extracting cells from an existing mesh using tria_generateSubdomain. A subdomain is always a set of cells, where some vertices/edges/faces may not be located at the physical boundary. These vertices/edges/faces are collected in an additional 'blind' boundary component, which receives number NBCT+1. In 2D, The subdomain is still related to the original boundary definition and does not have something like a 'local' t_boundary-object attached. The parameter values for vertices and edges on the blind boundary component are forced to be -1, as these actually inner-domain objects have no relationship to the parametrisation of the physical boundary.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@3077 4b12f95d-9313-0410-afc9-bd26427041f1