New features        : - Introduced the concept of 'extended raw' meshes. Problems solved     : Bugzilla-Bug-ID     : Additional comments :
When reading TRI files, all meshes are now converted to 'extended raw' meshes. These meshes provide a numbering for edges and faces additional to the vertex numbering. However, an optional parameter in the read routines allow to still read basic raw meshes, and the initStandardMeshFromRaw will also still work for this kind of meshes. Nevertheless, extended raw meshes will be necessary to provide additional information to the refinement routine is 2-level refinement for edges/faces will be introduced.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@3428 4b12f95d-9313-0410-afc9-bd26427041f1