New features        : Problems solved     : fixed: mlop_create2LvlMatrixStruct did not allocate Kdiagonal which led to strange errors... Bugzilla-Bug-ID     : Additional comments :
There is still one special case of type-9 matrices that we definately need to think about:
What happens if the matrix is non-square or does not have a main-diagonal for whatever reasons? Do we prescribe that Kdiagonal still has to be allocated (although senseless) - or do we say that Kdiagonal only exists if the matrix actually has a main diagonal?
Currently, the lsyssc_duplicateMatrix routine copies the handles from a source matrix to a destination matrix if LSYSSC_DUP_SHARE is specified, however, it does not set the LSYSSC_MSPEC_STRUCTUREISCOPY flag in the destination matrix if Kdiagonal does not exist in the source matrix - which leads to a "double-release" of the same handles once the source and destination matrices are destroyed - as happened in poisson2d_method1_l2prj.
What makes me wonder is that it seems like for some reason until now no one noticed that our poisson-application crashes - or at least no one complained about it... o_O

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@4046 4b12f95d-9313-0410-afc9-bd26427041f1