Implemented a new feature for both configure and the generated GNUMakefile: % ./configure --listids and % make listids now support wildcards in those matching rules defined in Makefile.cpu.inc.
Previously, no wildcard in the first three tokens of a build ID were feasible.
But to support all dialects of cygwin_nt* large lists of all possible cygwin_nt*
combinations would have been necessary.

Note that it is now feasible to define a matching rule in Makefile.cpu.inc like

   ifeq ($(call match,$(ID),pc-(coreduo|athlonxp)-cygwin_nt.*-g95-.*),yes)

as opposed to the previously necessary duplicate definitions

   ifeq ($(call match,$(ID),pc-coreduo-cygwin_nt.*-g95-.*),yes)
   ifeq ($(call match,$(ID),pc-athlonxp-cygwin_nt.*-g95-.*),yes)

But, please keep it simple in Makefile.cpu.inc as not every Featflow2 user
understands complex regular expressions as easily as do some people.
Typing 'make listids' and getting a 'list' of feasible build IDs that reads like

  (pc|pc64|ia64)-(core.*)-(linux|cygwin_nt.*)-g95-.*

might become a bit hard to read. Getting one of the following (depending on the
machine you're logged in to):

  pc-coreduo-(linux|cygwin_nt.*)-g95.*
  pc64-coreduo-(linux|cygwin_nt.*)-g95.*
  ia64-coreduo-(linux|cygwin_nt.*)-g95.*
  pc-coresolo-(linux|cygwin_nt.*)-g95.*
  pc64-coresolo-(linux|cygwin_nt.*)-g95.*
  ia64-coresolo-(linux|cygwin_nt.*)-g95.*

is preferable.



git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@4657 4b12f95d-9313-0410-afc9-bd26427041f1