Added FEAT2 regression benchmark. Using a copy of FLAGSHIP as first regression benchmark application and having defined exactly 1 FLAGSHIP test as benchmark test ID so far.
Usage:
  # to compile
  cd benchmark
  ./configure
  make benchmark

  # to run
  make alltests && ./runtests
or
  make alltests run

to see what's happening while benchmark is run use

  make alltests run SUPPRESSOUTPUT=0

Tests to be run are defined with the test ID in 'alltests.fbconf' and for
every test ID there is a test definition in tests/*.fbconf. More details
only on request (or look at the section 'FEAST regression benchmark' in
the FEAST documentation).


git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@4827 4b12f95d-9313-0410-afc9-bd26427041f1