Added CC2d as benchmark app. As source files use directly those stored in Featflow2/applications/cc2d/src.
There are a few things to be done yet for Michael:
Just grep for TODO_MICHAEL_KOESTER. In particular, the app is not yet
ready for being executed concurrently in a queueing system. Every output
file must be unique. It does not suffice to use some `date` or `pid` output.
In FEAST, we use the generic $LOGDIR env variable in master.dat's to have
unique path names. $LOGDIR is automatically set when generating the benchmark
control script 'runtests' and given a value according to the scheme 'logs/<test id>/'.

Strangely enough, the code seems to contain a bug in addition. To reproduce

  cd benchmark
  ./configure
  make feat2Benchmark-cc2d
  echo CC2D_001 > cc2d_problem.fbconf
  make cc2d_problem run

and you'll see the error
..../benchmark/feat2Benchmark-cc2d data/apps_cc2d/master.dat
* Warning (parlst_readfromfile): Specified data-subfile does not exist: ./data/output.dat
forrtl: severe (24): end-of-file during read, unit -5, file Internal List-Directed Read

although master.dat refers to a data/apps_cc2d/output.dat and nowhereelse "./data/output.dat" is refered to.


git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@4861 4b12f95d-9313-0410-afc9-bd26427041f1