BUG: Only the result of the program with highest refinement level used to be stored as reference solution.
But the intention is to store and compare results from all refinement levels.
     (Which is why it is considered bad and lazy practice to simply store the
      complete screen output. It would be preferable if every application would
      generate only 1 line with key values per program run. Such a line is likely
      to change less often than the complete screen output leading to a reduced
      frequency of benchmark result calibration.)

BUG: Detection of numactl binary and whether the kernel has NUMA features, too, was broken.

BUG: Finally found a way to deal with SUN's echo not being able to detect "-n" as command line option, but printing it to the screen.

ENH: In case KEEPGOING is set to 1 and some program runs failed, the results for all refinement levels that were successful are shown now.

     (Referred to as fragmented results.)

ENH: It is more natural to compare reference with current solution, not vice versa

ENH: Added support for Pathscale Debugger: make single DEBUGGER=pathdb

ENH: (Portably) Support benchmark binaries that are symbolic links

ENH: Avoid result file copy operation if that file was not created due to test failure

ENH: added an example of hostfiles/rankfiles on our GPU cluster that works perfectly well for Dom

TEXT: Fixed documentation on rankfiles (it was exactly the other way round on our GPU cluster)

TEXT: Adapted script to the fact that the configure command line option --mode and the Makefile
      variable MODE are deprecated and and have been replaced by --[no-]mpi and MPI=YES|NO.

This is a backport from FEAST.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@6971 4b12f95d-9313-0410-afc9-bd26427041f1