New features        : -ADDED: "include" command to fbdef files.
If in a .fbdef file the command
"include filename"
appears, the specified file is included at that point into the .fbdef file. This allows to exclude e.g. default settings for a set of tests into a separate file. Using the include statement then reads the file at the specific point, leading to a complete parameter reset.


git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@7556 4b12f95d-9313-0410-afc9-bd26427041f1