Update of the pseudo-templatised data structures.
The previous templatisation attempt lead to problems with module dependencies, at least with the IFC12 compiler. The compiler missed to compile all specialised modules (e.g. stackDble.f90, stackInt.f90, stackSngl.f90) before compiling the wrapper module (e.g. stack.f90).
The solution to this problem was to move the module-end-module pair as well as the use statement into the specialised F90 files instead of including them from the generic headers.

I also had to include the "kernel/DataStructures" directory manually into the FPP include directory list, as the IFC12 did not find the headers. This is quite surprising as the headers were #include'd in "quoted"-notation rather than <bracket>-notation, so according to the IFC12 FPP documentation the compiler should have found the headers...

PS: Fortunately we switched to C++ for FEAST2, so we can hopefully avoid outsmarting compilers in such ways in the future ^_^

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@9900 4b12f95d-9313-0410-afc9-bd26427041f1