* FIXED: bug in point coordinates array in QPW4DCP1 L2_0 filter * ADDED: L2_0 filters for EL_DCQP1_2D and EL_DCQP2_2D filters
Note: These new filters are so inefficient that it's not even funny. However, implementing efficient filters for elements beyond Q0 or QP1 requires that the filter chain, which is used by the linear solver along with other components performing the filtering, has to be capable of storing precomputed data, which is not the case in the current implementation, with discrete boundary conditions being one notable exception. Therefore we will have to revise the whole filter implementation to support more generic space constraints in the future. However, this undertaking requires serious modification on both kernel and application level, therefore this task has been recessed for now.
The filters implemented in this commit are therefore to be seen as provisional solutions which should be replaced once the previously mentioned changes are being performed.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10152 4b12f95d-9313-0410-afc9-bd26427041f1