* ENH: spdp_projectToVertices and spdp_projectToCells for check for P1/Q1 resp. P0/Q0 discretisations to avoid unnecessary calculation * ADDED: ucd_addVectorByVertex/ucd_addVectorByElement * ADDED: ucd_addVectorFieldByVertex/ucd_addVectorFieldByElement
The new ucd routines take a scalar vector (or an array thereof) as a parameter and automatically project the represented FE function (or vector field) into the vertices resp. cell midpoints. This eliminates the need for the user to explicit think about whether the underlying FE space of a vector has to be projected or not.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10243 4b12f95d-9313-0410-afc9-bd26427041f1