ADDED: Visual Studio project files for 'librariesNEW'
Michael and I have decided that we are going to redesign the Visual Studio build system for FeatFlow2 in near future. As a first step, we are switching to the new libraries, including fancy auto-download, auto-unpack and auto-compile features as in our un!x make system. Batch-Files for auto-compilation are upcoming soon.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10402 4b12f95d-9313-0410-afc9-bd26427041f1