CHANGED: switched from 'full debug info' to 'line numbers only' in 'Release' config for Visual Studio
Note: The release mode still includes debug information as this info is needed when profiling applications. However, generating and linking full debug information led to a serious slowdown of the code (factor ~10). Fortunately, generating and linking line number information does not seem to have negative impact on the code performance and it is sufficient to get some useful information from the profiler.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10492 4b12f95d-9313-0410-afc9-bd26427041f1