CHG: Renumberin strategies completely revised. That stuff now works with a real renumbering strategy structure.
REMARK: Theoretically, it is also possible to sort the rows of a matrix differently to the columns (necessary for off-diagonal block matrices), but this has not yet been implemented.
REMARK: Feast mirror boundary conditions cannot be used with resorted vectors at the moment... but anyway, they are not used.

CHG: Many "'" are changed to '"'.
FIX: Many PRINT-commands replaced by proper OUTPUT_LINE.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10538 4b12f95d-9313-0410-afc9-bd26427041f1