DISABLED: automatic library archive unpacking for Windows
Using gzip under Windows leads to problems with file access rights on NTFS partitions.
In particular, even with administrator rights, the user may not be allowed to delete files created by gzip, thus leading to persistent files which can only be killed by manually adjust access rights. This behavior is non-deterministic; it may or may not screw up file rights.

This problem has been known to me from using cygwin as a build environment on NTFS partitions. Most likely gzip uses some faulty code parts from cygwin for the interaction with the file system.

In consequence, the automatic unpacking of downloaded library archives has been disabled and - as screwing up the users file system poses a not-to-be-underestimated hazard - the gzip.exe and tar.exe binaries have been deleted.

We will have to find another way of providing an auto-unpack feature for the libraries - an open-source Windows-native unpacker would be most desirable for that issue.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10874 4b12f95d-9313-0410-afc9-bd26427041f1