FIX: IFC13 not supported.
NOTE: There is a problem with IFC13. The project file has to be converted, i.e., it has to be opened once in the IDE, be converted and saved again. Otherwise, most of the modules would not compile -- and they don't give an error message about that.

We have to revise this...

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@10936 4b12f95d-9313-0410-afc9-bd26427041f1