CHG: Some code cleanup.
This code uses the approach u=P(-1/alpha lambda). In the end, the code does not give the behaviour promised. Example: Optimal distributed control of Driven Cavity. Without bounds, it looks ok. With bounds enabled, for a short time, it converges as well. For a longer time however, the dual solution shows that much of a weird behaviour, that the Newton iteration does not converge. Although the CG algorithm always converges fine, the nonlinear solution always switches between two states without converging. This completely contradicts the experiences of other authors [Ulbrich], where the solver converges fine. So code development is postponed.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@11616 4b12f95d-9313-0410-afc9-bd26427041f1