CHG: Some code cleanup.
This code uses the approach u=P( (1-alpha) u + lambda)=P(u-sigma(alpha u+lambda)) with sigma=1. The iteration takes place directly in the control space, residuals do not contain mass matrices! All in all, the approach seems to work fine for Dirichlet and L2 boundary control.

The implementation of H^(1/2) boundary control is completely wrong! The code uses the L2 projection on the boundary of the dual variable to the control space; actually, there would have an H^(1/2) projection to be used which is nonlocal! So all results generated with this code using H^(1/2) boundary control are wrong.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@11617 4b12f95d-9313-0410-afc9-bd26427041f1