added caller string to storage node
The name of the routine allocating a storage node has been passed to the storage since the beginning, however, it has never been stored or processed in any way.
Now the storage node saves the name of the calling routine and also prints this name in a storage_info call, which should slightly ease the pain when searching for memory leaks.

git-svn-id: svn://doncamillo.mathematik.tu-dortmund.de/Featflow2@11698 4b12f95d-9313-0410-afc9-bd26427041f1