Deep modifications of the extended finite element comparer, many under the hood Added support for the output of flagship Added support for 1D and created options to easily add 3D Added support for discretisations with only 1 element type and not only element pairs Added options to easily add more types of discretisations
Slightly extended the options what you can calculate

Created the possibility to add more features - at the moment this tool can only calculate L2-Norms/L2-Errors/Pointvalues/Differences in the point values of finite element functions.
All it needs as input is: A vector with the coefficents, the .prm and .tri-file and the level of the solution. Note that all that is neccesary is some sort of overlap of the meshes to be able to use the tool.

Deleted all data that is not needed to run this tool.



git-svn-id: https://svn.mathematik.tu-dortmund.de:20443/svn/Featflow2@12055 4f116ff2-4e46-254c-a92c-0a7323bc4f3a