added checks whether the deformation tensor can actually be assembled
If one wants to assemble the deformation tensor, it is necessary to set the "iUpwind" parameter of [CC-DISCRETISATION] to either 4 or 5, as the other (old) assembly routines do not support the deformation tensor at all. This was already documented in the discretisation.dat file, however, the cc2d code did not check this.
In consequence, if one had set isubequation=1 and iUpwind=0, the code silently assembled a null matrix without emitting any warning or error message.

This commit introduces a few additional checks to ensure that the user selected an appropriate assembly method in case of deformation tensor formulation.
Also clarified the error message emitted when trying to assemble non-constant viscosity, which is also only supported by iUpwind \in {4,5}.

git-svn-id: https://svn.mathematik.tu-dortmund.de:20443/svn/Featflow2@12215 4f116ff2-4e46-254c-a92c-0a7323bc4f3a