provide real point coordinates for non-constant viscosity SD assembly
In case of non-constant viscosity models, the cc2d (or more precisely: the underlying streamline-diffusion assembly routine) did not provide the domain point-coordinates to the callback function, which made it impossible to prescribe viscosity functions depending on the point coordinates.

Now, the SD assembly method always provides real point coordinates if the viscosity is non-constant.
This may lead to slightly longer assembly times when running cc2d with non-constant viscosity, however, the impact should be negligible.

git-svn-id: https://svn.mathematik.tu-dortmund.de:20443/svn/Featflow2@12216 4f116ff2-4e46-254c-a92c-0a7323bc4f3a