fixed "new" Vanka Halloween bug
There was a typo in the "new" implementation of the "full" Vanka smoother for the Q1~/Q0 and P1~/P0 elements, which was introduced in rev. 3887 (submitted on 31.10.2008 by myself...).

Note:
Unless you have been using "ismootherType=102" in combination with the Q1~/Q0 element in the cc2d application, this commit should not affect your simulations and/or previously computed results.

git-svn-id: https://svn.mathematik.tu-dortmund.de:20443/svn/Featflow2@12222 4f116ff2-4e46-254c-a92c-0a7323bc4f3a