FIX: always download BLAS-3.5.0 tarball
Until now, the most up-to-date tarball of the BLAS library was downloaded, which led to version conflicts between the tarball's contents and our own custom makefile. To avoid further version conflicts, the configure script now always downloads version 3.5.0 of the BLAS library even if a newer version is available. This ensures that we do not have to update our scripts every time a new BLAS version is released.

git-svn-id: https://svn.mathematik.tu-dortmund.de:20443/svn/Featflow2@12252 4f116ff2-4e46-254c-a92c-0a7323bc4f3a